### Description

Suggestion by @Tyrrrz to improve performance was actually much slower: https://github.com/bitwarden/passwordless-server/pull/518

### Shape

n/a

### Screenshots

| Method | Mean     | Error   | StdDev  | Gen0   | Gen1   | Allocated |
|------- |---------:|--------:|--------:|-------:|-------:|----------:|
| New    | 139.4 ns | 0.77 ns | 0.64 ns | 0.3068 | 0.0012 |    2568 B |
| Old    | 101.5 ns | 1.22 ns | 1.14 ns | 0.1013 |      - |     848 B |

The old one is the change I'd want to merge back into `main`.

### Checklist
I did the following to ensure that my changes were tested thoroughly:
- __

I did the following to ensure that my changes do not introduce security vulnerabilities:
- __